refactor: rename and replace path
